Regionais/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
pydantic==1.10.7
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
kafka-python==2.0.2
python-dotenv==1.0.0

